providers/anyscale/hooks/anyscale.py
from airflow.hooks.base import BaseHook
from anyscale import AnyscaleSDK

class AnyscaleHook(BaseHook):
    def __init__(self, anyscale_conn_id='anyscale'):
        super().__init__()
        self.anyscale_conn_id = anyscale_conn_id
        self.sdk = None

    def get_conn(self):
        """
        Authenticate and return an Anyscale SDK client.
        """
        if self.sdk is not None:
            return self.sdk

        # Fetch connection details from Airflow
        conn = self.get_connection(self.anyscale_conn_id)

        # Use the connection's password field to store the auth token
        auth_token = conn.get_password()

        # Initialize the Anyscale SDK client
        self.sdk = AnyscaleSDK(auth_token=auth_token)

        return self.sdk

    def create_cloud(self):
        pass

    def delete_cloud(self):
        pass

    def create_compute_config(self):
        pass

    def create_cluster(self):
        pass

    def delete_cluster(self):
        pass

    def submit_anyscale_job(self):
        pass

    def cancel_anyscale_job(self):
        pass

    def rollout_anyscale_service(self):
        pass

    def delete_anyscale_service(self):
        pass

    def launch_cluster(self, project_id, cluster_name, cluster_environment_build_id, cluster_compute_id=None, **kwargs):
        """
        Launch a cluster using the Anyscale SDK.
        """
        sdk = self.get_conn()
        cluster = sdk.launch_cluster(
            project_id=project_id,
            cluster_name=cluster_name,
            cluster_environment_build_id=cluster_environment_build_id,
            cluster_compute_id=cluster_compute_id,
            **kwargs
        )
        return cluster
